==> notebooks/1_train_test_val_split.ipynb <==
# Load dataset


from pathlib import Path

base_path = Path(".").absolute().parents[0]

base_path

import sys

sys.path.insert(0, str(base_path / "src"))

str(base_path / "src")

from google.cloud import bigquery

client = bigquery.Client(project="artefact-taxonomy")


from utils.text_utils import clean_html
from sklearn.model_selection import train_test_split

CUTOFF = 5

# Define the product_id and target / label
id_col = "adeo_product_id"
title_col = "title"
language_col = "lang"
description_raw = "description"
description_col = "description_clean"
label_raw_col = "mirakl_model_code"

import pandas as pd

pd.options.mode.chained_assignment = None  # default='warn'

sql = """
SELECT * FROM `artefact-taxonomy.pem_uc_add_datasets.temp_simplon_2`
"""

df = client.query(sql).to_dataframe()

df.to_csv('../data/full_data.csv', index=False)

df = pd.read_csv('../data/full_data.csv')

## Check the data
- 100 mirakl classes
- with 4 other attributes:
    style: 02419
    color: 01746
    shape: 00562
    battery included: 15344

df.info()

### Number of class available

len(df['mirakl_model_code'].unique())

df.groupby('attribute_code').count()

### Check if any null values

df.isnull().sum()

df.isnull().values.any()

# Data Cleaning

df[description_col] = df[description_raw].copy
df[description_col] = (
    df[description_raw].loc[df[description_raw].notnull()].apply(clean_html)
)
df.fillna("", inplace=True)

df.isnull().values.any()

## Preprocessing

### Down sampling to have only 500 samples (mixture of attributes) per class

n = 500
msk = df.groupby(label_raw_col)[label_raw_col].transform('size') >= n
df_cut = pd.concat((df[msk].groupby(label_raw_col).sample(n=n), df[~msk]), ignore_index=True)
print(df_cut.shape)

### Get unique id to ensure a single product will only be either in train or test or val (avoid data leaking)

# Get unique id to ensure a single product will only be either in train or test or val 
# To avoid data leaking
# df_unique = df_cut.copy()
# df_unique.drop(columns=[title_col, language_col, description_raw], inplace=True)
df_unique = df_cut[[id_col,label_raw_col]]
df_unique = df_unique.drop_duplicates(subset=id_col, keep="first")

# Calculate the no. of entries and models removed due to sample <= CUTOFF per model
sample_cut_lackofsample = df.shape[0] - df_cut.shape[0]
model_cut_lackofsample = df[label_raw_col].nunique() - df_cut[label_raw_col].nunique()

# Calculate the no. of entries and models removed due to sample <= CUTOFF per model
sample_cut_lackofinfo = df.shape[0] - df_cut.shape[0] - sample_cut_lackofsample
model_cut_lackofinfo = (
    df[label_raw_col].nunique() - df_cut[label_raw_col].nunique() - model_cut_lackofsample
)

# Split the unique adeo_product_id
# To ensure that a unique adeo_product_id exist in only 1 dataset - to avoid data leaking

train_val, test = train_test_split(df_unique, test_size=0.1, stratify=df_unique[label_raw_col])
train, val = train_test_split(train_val, test_size=0.1, stratify=train_val[label_raw_col])

val.columns

df_cut.columns

## Merge it back to the main df 

del val[label_raw_col]
val = val.merge(df_cut, how="inner", on=id_col)

del train[label_raw_col]
train = train.merge(df_cut, how="inner", on=id_col)

del test[label_raw_col]
test = test.merge(df_cut, how="inner", on=id_col)

## Check for overlapping

import numpy as np


def check_if_overlap(df1, df2, on_column):
    # merge two dataFrames and add indicator column
    all_df = pd.merge(df1, df2, on=on_column, how="left", indicator="exists")
    # add column to show if each row in first DataFrame exists in second
    all_df["exists"] = np.where(all_df.exists == "both", True, False)
    if all_df["exists"].sum() == 0:
        print("These 2 dataframes are not overlapped over ", on_column)
    else:
        print("There are overlapped entries!!")

check_if_overlap(val, train, id_col)

check_if_overlap(test, train, id_col)

check_if_overlap(test, val, id_col)

## Summary

print("============================")
print("Full Dataset")
print("============================")
print("No. entries:", df.shape[0])
print("No. of models (categories / label): ", df[label_raw_col].nunique())

print("No. of entries removed due to num of samples < CUTOFF: ", sample_cut_lackofsample)
print("No. of models removed due to num of samples < CUTOFF: ", model_cut_lackofsample)

print(
    "No. of entries removed due to lack of product information (no title & no description): ",
    sample_cut_lackofinfo,
)
print(
    "No. of models removed due to to lack of product information (no title & no description): ",
    model_cut_lackofinfo,
)

print("No. of entries after cut ", (df_cut.shape[0]))
print("No. of models after cut ", (df_cut[label_raw_col].nunique()))

print("============================")
print("Train Dataset")
print("============================")
print("No. entries:", train.shape[0])
print("No. of models (categories / label): ", train[label_raw_col].nunique())

print("============================")
print("Val Dataset")
print("============================")
print("No. entries:", val.shape[0])
print("No. of models (categories / label): ", val[label_raw_col].nunique())

print("============================")
print("Test Dataset")
print("============================")
print("No. entries:", test.shape[0])
print("No. of models (categories / label): ", test[label_raw_col].nunique())


print("============================")
print("Verify if Dataset is correct")
print("============================")
if train.shape[0] + val.shape[0] + test.shape[0] == df_cut.shape[0]:
    print("The total no. of entries matched!!")
else:
    print("No. of entries matching ERROR!!")

## Save train, val, test split separately

train.to_csv('../data/train.csv', index=False)

val.to_csv('../data/val.csv', index=False)

test.to_csv('../data/test.csv', index=False)
