==> notebooks/3_training_classification.ipynb <==
# Import packages

import json
from functools import partial
from pathlib import Path

import pandas as pd
import pytorch_lightning
import torch
from torch.utils.data import DataLoader
from transformers import AutoTokenizer

from attribute_extraction.models.attribute_classification import MultiAttributeClassifier
from attribute_extraction.models.mapper import Mapper
from attribute_extraction.models.train_utils import (
    AttributeDataset,
    MultiAttributeBatchSampler,
    build_callbacks,
    collate_fun_generator,
)
from attribute_extraction.utils.data_balancing_utils import add_weight_for_data_balancing


# setting up variables and hyperparameters

attribute_code_col = "attribute_code"
attribute_lov_col = "lov_code"
context_col = "context"
weight_col = "weight"
mapped_lov_col = "mapped_lov_code"

local_path = "../outputs_train_workflow/"

model_name = "distilbert-base-multilingual-cased"
experiment_description = "Multi-task small GRU model with multilanguage distilledBert tokenizer"
num_epoch = 20
max_len = 512
batch_size = 128 
freeze_backbone = True
learning_rate = 1e-4
projection_dim = 256
dropout = 0.2
data_balance = False
upper_qn = 0.9
lower_qn = 0.1
lov_attribute_codes = ["02419", "01746", "00562", "15344", "99999"]
train_set_uri = "../data/train_formatted.csv"
val_set_uri = "../data/val_formatted.csv"
test_set_uri = "../data/test_formatted.csv"

# Save the hyper_parameters
hyper_parameters = {
    "model_name": model_name,
    "experiment_description": experiment_description,
    "num_epoch": num_epoch,
    "max_len": max_len,
    "batch_size": batch_size,
    "freeze_backbone": freeze_backbone,
    "learning_rate": learning_rate,
    "projection_dim": projection_dim,
    "dropout": dropout,
    "data_balance": data_balance,
    "upper_qn": upper_qn,
    "lower_qn": lower_qn,
    "lov_attribute_codes": lov_attribute_codes,
    "train_set_uri": train_set_uri,
    "val_set_uri": val_set_uri,
    "test_set_uri": test_set_uri,

}

with open(f"{local_path}hyper_parameters.json", "w") as f:
    json.dump(hyper_parameters, f, indent=4)

# Loading data

data_train = pd.read_csv(train_set_uri).reset_index(drop=True)
data_val = pd.read_csv(val_set_uri).reset_index(drop=True)
data_test = pd.read_csv(test_set_uri).reset_index(drop=True)

data_train['attribute_code'] = data_train['attribute_code'].apply(lambda x: str(x).zfill(5))
data_train['lov_code'] = data_train['lov_code'].apply(lambda x: str(x).zfill(5))

data_val['attribute_code'] = data_val['attribute_code'].apply(lambda x: str(x).zfill(5))
data_val['lov_code'] = data_val['lov_code'].apply(lambda x: str(x).zfill(5))

data_test['attribute_code'] = data_test['attribute_code'].apply(lambda x: str(x).zfill(5))
data_test['lov_code'] = data_test['lov_code'].apply(lambda x: str(x).zfill(5))

data_train = data_train.fillna('')
data_val = data_val.fillna('')
data_test = data_test.fillna('')

data_train[context_col] = data_train.title + " " + data_train.description_clean
data_val[context_col] = data_val.title + " " + data_val.description_clean
data_test[context_col] = data_test.title + " " + data_test.description_clean

data_train.isnull().sum()


# Building Mapper

mapper = Mapper(attribute_code_col=attribute_code_col, attribute_value_col=attribute_lov_col)

mapper.fit(pd.concat([data_train, data_val, data_test]))

mapper.save(f"{local_path}mapper.json")

# Mapping Columns

data_train = mapper.map_dataframe(data_train, mapped_col_name=mapped_lov_col)
data_val = mapper.map_dataframe(data_val, mapped_col_name=mapped_lov_col)
data_test = mapper.map_dataframe(data_test, mapped_col_name=mapped_lov_col)

# Data balancing

if data_balance:

    data_train = add_weight_for_data_balancing(
        df=data_train,
        label_col=mapped_lov_col,
        weight_col=weight_col,
        attribute_code_col=attribute_code_col,
        upper_qn=upper_qn,
        lower_qn=lower_qn,
    )

    data_val[weight_col] = 1
    data_test[weight_col] = 1
else:
    data_train[weight_col] = 1
    data_val[weight_col] = 1
    data_test[weight_col] = 1

# Model Initialisation

tokenizer = AutoTokenizer.from_pretrained(model_name)
model = MultiAttributeClassifier(
    vocab_size=tokenizer.vocab_size,
    class_config=mapper.mappings,
    freeze_backbone=freeze_backbone,
    warmup_steps=data_train.shape[0] // batch_size // 100,
    estimated_stepping_batches=data_train.shape[0]
    * num_epoch
    // batch_size,
    num_cycles=num_epoch,
    projection_dim=projection_dim,
    dropout=dropout,
    learning_rate=learning_rate,
)

# Dataloaders Initialisation

train = AttributeDataset(
    data=data_train,
    context_col_name=context_col,
    label_col_name=mapped_lov_col,
    attribute_code_col_name=attribute_code_col,
)
validation = AttributeDataset(
    data=data_val,
    context_col_name=context_col,
    label_col_name=mapped_lov_col,
    attribute_code_col_name=attribute_code_col,
)
test = AttributeDataset(
    data=data_test,
    context_col_name=context_col,
    label_col_name=mapped_lov_col,
    attribute_code_col_name=attribute_code_col,
)

train_loader = DataLoader(
    dataset=train,
    batch_sampler=MultiAttributeBatchSampler(
        data=data_train,
        batch_size=batch_size,
        split_col=attribute_code_col,
        weight_col=weight_col,
    ),
    collate_fn=partial(
        collate_fun_generator,
        tokenizer=tokenizer,
        max_len=max_len,
    ),
    num_workers=4,
)

validation_loader = DataLoader(
    dataset=validation,
    batch_sampler=MultiAttributeBatchSampler(
        data=data_val,
        batch_size=batch_size,
        split_col=attribute_code_col,
        weight_col=weight_col,
    ),
    collate_fn=partial(
        collate_fun_generator,
        tokenizer=tokenizer,
        max_len=max_len,
    ),
    num_workers=4,
)

test_loader = DataLoader(
    dataset=test,
    batch_sampler=MultiAttributeBatchSampler(
        data=data_test,
        batch_size=batch_size,
        split_col=attribute_code_col,
        weight_col=weight_col,
    ),
    collate_fn=partial(
        collate_fun_generator,
        tokenizer=tokenizer,
        max_len=max_len,
    ),
    num_workers=4,
)

train.data

# Build Callbacks

callbacks, metric_logger = build_callbacks(output_path=Path(local_path), model_name="model")

# Model Training

trainer = pytorch_lightning.Trainer(
    gpus=1 if torch.cuda.is_available() else 0,
    logger=metric_logger,
    callbacks=callbacks,
    log_every_n_steps=1,
    max_epochs=num_epoch,
    precision=16 if torch.cuda.is_available() else 32,
)

trainer.fit(model, train_loader, validation_loader)

# Model Evaluation

test_metrics = trainer.test(model, test_loader)
test_metrics_sort = [dict(sorted(test_metrics[0].items()))]

with open(local_path / "metrics.json", "w") as f:
    json.dump(test_metrics, f, indent=4)

with open(local_path / "metrics_sorted.json", "w") as f:
    json.dump(test_metrics_sort, f, indent=4)
