==> notebooks/6_evaluation_classification.ipynb <==
# import package

from pathlib import Path

import pandas as pd

from attribute_extraction.utils.att_lov_extract import (
    get_classification_report,
    get_performance_metrics_table,
)

# Setting up variables

attribute_code_col = "attribute_code"
attribute_label_col = "attribute_label"
attribute_lov_col = "lov_code"
language_col = "lang"

pred_label_col = "predicted_labels"

predictions_path = "../outputs_predict_workflow/predictions.csv"


# Loading Predictions data


predictions = pd.read_csv(predictions_path).reset_index(drop=True)

predictions

attribute_codes = predictions[attribute_code_col].unique()
attribute_codes.sort()

df_print_all = pd.DataFrame(
    columns=[
        attribute_code_col,
        attribute_label_col,
        "Dataframe",
        "Accuracy (%)",
        "Precision (Weighted-avg)(%)",
        "Precision (Macro-avg) (%)",
        "Recall (Weighted-avg) (%)",
        "Recall (Macro - avg) (%)",
        "Number of samples",
    ]
)

for attribute_code in attribute_codes:

    attribute_label = (
        predictions[attribute_label_col]
        .loc[
            (predictions[attribute_code_col] == attribute_code)
            & (predictions[language_col] == "FR")
        ]
        .values[0]
    )

    attribute_data = predictions[predictions[attribute_code_col] == attribute_code]

    df_metric_overall, df_metric_detail = get_classification_report(
        attribute_data[attribute_lov_col], attribute_data[pred_label_col]
    )

    d_eval = {
        "Overall": df_metric_overall,
    }

    for language in ["FR", "ES", "IT"]:

        df_eval_lang = attribute_data.loc[attribute_data[language_col] == language]

        df_metric_overall_lang, df_metric_detail_lang = get_classification_report(
            df_eval_lang[attribute_lov_col], df_eval_lang[pred_label_col]
        )

        d_eval[language] = df_metric_overall_lang

    df_print = get_performance_metrics_table(d_eval)

    df_print["attribute_code"] = str(attribute_code).zfill(5)
    df_print["attribute_label"] = attribute_label

    df_print_all = pd.concat([df_print_all, df_print])



df_print_all
